+++ src/components/DesktopNavbar.jsx
@@ -4,10 +4,8 @@
 // #0058ea light blue
 
 import { Link, useLocation } from "react-router-dom"
-import RiTwitterXFill from "./Icons/RiTwitterXFill"
-import MdiInstagram from "./Icons/MdiInstagram"
-import HugeiconsFacebook02 from "./Icons/HugeiconsFacebook02"
-import PhLinkedinLogo from "./Icons/PhLinkedinLogo"
+import { FaFacebookF, FaLinkedinIn } from "react-icons/fa"
+import { RiInstagramFill } from "react-icons/ri";
 
 const Navbar = () => {
   const location = useLocation()
@@ -18,32 +16,12 @@ const Navbar = () => {
       name: "Overview"
     },
     {
-      linkTo: "/about-us/history",
-      name: "History"
-    },
-    {
-      linkTo: "/about-us/clients-and-testimonials",
-      name: "Clients & Testimonials"
-    },
-    {
-      linkTo: "/about-us/personal-biography",
-      name: "Personal biography"
-    },
-    {
       linkTo: "/about-us/team-members",
       name: "Team Members"
     },
     {
-      linkTo: "/about-us/location",
-      name: "Location"
-    },
-    {
       linkTo: "/about-us/contact",
       name: "Contact"
-    },
-    {
-      linkTo: "/about-us/construction",
-      name: "Construction"
     }
   ]
   return (
@@ -59,25 +37,22 @@ const Navbar = () => {
             )}
           </div>
         </li>
-        <li className={`nav px-3 pb-3 mt-4 relative ${location.pathname === "/services" ? "bg-white text-[navy] font-semibold" : "text-[whitesmoke]"}`}>
-          <Link to="/services">Services</Link>
-
-          <div className="subnav absolute hidden z-50">
-            <Link to="/services" className="p-[10px] border-b">Overview</Link>
-            <Link to="/services/faqs" className="p-[10px] border-b">FAQ</Link>
-          </div>
-        </li>
-        <li className={`px-3 pb-3 mt-4 ${location.pathname === "/projects" ? "bg-white text-[navy] font-semibold" : "text-[whitesmoke]"}`}><Link to="/projects">Projects</Link></li>
+        
         <li className={`px-3 pb-3 mt-4 ${location.pathname === "/blogs" ? "bg-white text-[navy] font-semibold" : "text-[whitesmoke]"}`}><Link to="/blogs">Blog/News</Link></li>
       </ul>
 
       <div className="flex items-center gap-4">
-        <a href=""><RiTwitterXFill className="text-[whitesmoke] text-xl"/></a>
-        <a href=""><MdiInstagram className="text-[whitesmoke] text-xl"/></a>
-        <a href=""><HugeiconsFacebook02 className="text-[whitesmoke] text-xl"/></a>
-        <a href=""><PhLinkedinLogo className="text-[whitesmoke] text-xl"/></a>
+        <a href="https://www.facebook.com/share/15oEKg5tux/" target="_blank">
+          <FaFacebookF className="text-[whitesmoke] text-xl"/>
+        </a>
+        <a href="https://www.instagram.com/ictopia1?igsh=MXY0cTRlMGVreGtjMw==" target="_blank">
+          <RiInstagramFill className="text-[whitesmoke] text-xl"/>
+        </a>
+        <a href="https://www.linkedin.com/company/ictopia/" target="_blank">
+          <FaLinkedinIn className="text-[whitesmoke] text-xl"/>
+        </a>
 
-        <button className="bg-[#54c4cf] hover:text-[navy] transition-all duration-1000 py-[3px] px-3 text-[whitesmoke] text-sm font-semibold"><Link to="/register">Get a quote</Link></button>
+        <button className="bg-[#54c4cf] hover:text-[navy] transition-all duration-1000 py-[3px] px-3 text-[whitesmoke] text-sm font-semibold rounded-md"><Link to="/register">Register</Link></button>
       </div>
     </nav>
   )
