--- src/components/ProjectsComponents/Header.jsx
@@ -1,8 +1,0 @@
-
-const Header = () => {
-  return (
-    <div>Projects</div>
-  )
-}
-
-export default Header
