--- src/views/Projects.jsx
@@ -1,11 +1,0 @@
-import Header from "../components/ProjectsComponents/Header"
-
-const Projects = () => {
-  return (
-    <main>
-        <Header />
-    </main>
-  )
-}
-
-export default Projects
