--- src/views/Service/Faq.jsx
@@ -1,13 +1,0 @@
-import Footer from "../../components/Footer"
-import FaqQ from "../../components/ServicesComponents/FaqQ"
-
-
-const Faq = () => {
-  return (
-    <div>
-      <FaqQ/>
-    </div>
-  )
-}
-
-export default Faq;
