--- src/views/Service/Overview.jsx
@@ -1,15 +1,0 @@
-import IndustriesServed from "../../components/ServicesComponents/IndustriesServed"
-import CompanyService from "./CompanyService"
-import ServiceSection from "../../components/ServicesComponents/ServiceSection"
-
-const Overview = () => {
-  return (
-    <div>
-      <ServiceSection />
-      <CompanyService />
-      <IndustriesServed />      
-    </div>
-  )
-}
-
-export default Overview
