--- src/views/Service/Services.jsx
@@ -1,12 +1,0 @@
-import Overview from "./Overview"
-
-
-const Services = () => {
-  return (
-    <main>
-        <Overview />
-    </main>
-  )
-}
-
-export default Services
